Follow-up commit (parse playbackSpeed from launch files)

+++ DSO_ROS/catkin_ws/src/dso_ros/include/dso_ros/dso_node.h
@@ -114,7 +114,7 @@ public:
 
   int original_setting_photometricCalibration;
 
-  float playbackSpeed;
+  double playbackSpeed;
   int nSkippedFrames;
   int nTotalFramesAfterInit;
   std::vector<double> trackingTimes; //in ms
